Fixes and improvements result of the feedback of the Deliverable 2
Moving the styled components into their own files
Renaming files to specify the content
Creating sub-folder to group common files
