+++ src/App.js
@@ -1,26 +1,9 @@
 import { useState } from 'react';
-import styled from 'styled-components';
-import './App.css';
-import Header from './components/Header/Header.js';
-import Home from './components/Home';
-import ProductList from './components/ProductList';
+import { Content, Footer, Section } from './App.styles';
+import Header from './components/Header/Header.component';
+import Home from './components/Home.component';
+import ProductList from './components/ProductList/ProductList.component';
 import { useFeaturedBanners } from './utils/hooks/useFeaturedBanners';
-
-const Section = styled.section`
-  max-width: 960px;
-  margin: auto;
-`;
-
-const Content = styled.section`
-  padding: 10px;
-`;
-
-const Footer = styled.section`
-  text-align: center;
-  padding-bottom: 20px;
-  padding-top: 10px;
-  border-top: 1px solid lightgray;
-`;
 
 function App() {
   const { data, isLoading } = useFeaturedBanners();
