+++ src/App.styles.jsx
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+const Section = styled.section`
+max-width: 960px;
+margin: auto;
+`;
+
+const Content = styled.section`
+padding: 10px;
+`;
+
+const Footer = styled.section`
+text-align: center;
+padding-bottom: 20px;
+padding-top: 10px;
+border-top: 1px solid lightgray;
+`;
+
+export { Section, Content, Footer };
