--- src/components/Header/ImgRounded.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const ImgRounded = styled.img`
-  border-radius: 50%;
-  width: 50px;
-`;
-
-export default ImgRounded;
