--- src/components/Header/SearchCartContainer.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const SearchCartContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-`;
-
-export default SearchCartContainer;
