--- src/components/Header/StyledHeader.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-const StyledHeader = styled.header`
-  display: flex;
-  align-items: center;
-  border-bottom: 1px solid lightgray;
-`;
-
-export default StyledHeader;
