+++ src/components/Header/Title.component.js
@@ -1,0 +1,9 @@
+import { TitleContainer, TitleLink } from './Title.styles';
+
+const Title = ({title, handleGoToPage}) => (
+  <TitleContainer>
+    <TitleLink alt={title} onClick={handleGoToPage} data-go-to-page="Home">{title}</TitleLink>
+  </TitleContainer>
+);
+
+export default Title;
