+++ src/components/Header/Title.styles.jsx
@@ -14,10 +14,4 @@ const TitleLink = styled.a`
   cursor: pointer;
 `;
 
-const Title = ({title, handleGoToPage}) => (
-  <TitleContainer>
-    <TitleLink alt={title} onClick={handleGoToPage} data-go-to-page="Home">{title}</TitleLink>
-  </TitleContainer>
-);
-
-export default Title;
+export { TitleContainer, TitleLink };
