+++ src/components/Product.component.js
@@ -1,4 +1,4 @@
-import { Card } from './Card';
+import { Card } from './Shared/Card/Card.component';
 
 export default function Product({product}) {
     return (
