+++ src/components/Slider/Slider.styles.jsx
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import items from '../mocks/en-us/featured-banners.json';
 
 const SliderContainer = styled.div`
   width: 100%;
@@ -24,17 +23,4 @@ const SliderTitle = styled.div`
   font-weight: bold;
 `;
 
-export default function Slider() {
-  const banners = items.results;
-
-  return (
-    <SliderContainer>
-      {
-        banners.map(({id, data, href}) =>
-          <SliderItem key={id} image_url={data.main_image.url} >
-            <SliderTitle href={href} alt={data.main_image.alt}>{data.title}</SliderTitle>
-          </SliderItem>)
-      }
-    </SliderContainer>
-  );
-}
+export { SliderContainer, SliderItem, SliderTitle };
